% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{actorStat}
\alias{actorStat}
\title{actorStat}
\usage{
actorStat(values, type, edgelist, riskset)
}
\arguments{
\item{values}{3-column matrix (id, change time, value). Actor ids should 
correspond to actor ids in the edgelist and riskset. Change time is zero 
for all entries if the covariate is time invariate.}

\item{type}{1 = sender effect, 2 = receiver effect}

\item{edgelist}{3-column edgelist (time, sender, receiver)}

\item{riskset}{2-column riskset (sender/actor 1, receiver/actor 2).}
}
\value{
matrix (time x dyad)
}
\description{
A function to transform exogenous actor covariate variables in the format ..
requested for estimation with relevent::rem().
}
\examples{
data(edgelistD)
data(covar)
out <- prepER(edgelist = edgelistD)
el <- out$edgelist
rs <- out$rs
ac <- out$ac
covar$id <- ac$id[match(covar$id, ac$name)]
covar <- as.matrix(covar)
stat <- actorStat(values = covar[,c(1:3)], type = 1, edgelist = el, 
    riskset = rs)

}
