% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{baseline}
\alias{baseline}
\alias{intercept}
\title{baseline}
\description{
A baseline effect in the \code{effects} argument of \code{\link{tomstats}} 
is automatically specified when \code{ordinal} in \code{\link{tomstats}} is 
set to FALSE (default) and automatically removed when this argument is set 
to TRUE. Alternatively, a baseline effect can be explicitly specified by 
adding '1' to the equation or explicitly removed by adding '-1' to the 
equation. This also holds for the \code{rateEffects} in 
\code{\link{aomstats}}.
}
\details{
The baseline effect refers to the baseline tendency to interact. In the 
tie-oriented model, the log-inverse of the estimated parameter translates to 
the average number of observed events per time unit per dyad. In the 
actor-oriented model, the log-inverse of the estimated parameter translates 
to the average number of observed events per time unit per actor. The 
statistic is equal to one for all dyads resp. actors in the riskset at all 
timepoints.
}
\examples{
tomstats(~ 1, edgelist = history)
aomstats(rateEffects = ~ 1, edgelist = history)
}
