% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{both_male_long}
\alias{both_male_long}
\title{Exogenous Dyad Attribute in Long Format: both_male_long}
\format{
A data frame with the following columns:
\describe{
 \item{actor1}{Numeric id of the first actor in the dyad.}
 \item{actor2}{Numeric id of the second actor in the dyad.}
 \item{both_male}{Binary indicator (1 for male-male dyads, 0 otherwise).}
}
}
\source{
Simulated exogenous information on actors in a social network.
}
\usage{
data(both_male_long)
}
\description{
A data frame representing exogenous attributes of dyads in a social network in long format. Each row indicates whether a dyad consists of two male actors (sex=0) in the original matrix `info_both_male_wide`.
}
\examples{
data(both_male_long)
head(both_male_long)

}
\seealso{
\code{\link{calculate_tie_statistic}} for the function using this data, \code{\link{both_male_wide}} for the data in wide format, and \code{\link{info}} for an overview of the actor exogenous attributes.
}
\keyword{dataset}
