% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{both_male_wide}
\alias{both_male_wide}
\title{Exogenous Dyad Attribute Matrix: both_male_wide}
\format{
A square matrix with dimensions equal to the number of unique actors.
}
\source{
Simulated exogenous information on actors in a social network.
}
\usage{
data(both_male_wide)
}
\description{
A matrix representing exogenous attributes of dyads in a social network.
The matrix indicates whether a dyad consists of two male actors (sex=0).
Rows and columns correspond to actor IDs, and cells contain binary values
(1 for male-male dyads, 0 otherwise).
}
\examples{
data(both_male_wide)
print(both_male_wide)

}
\seealso{
\code{\link{tie}} for the function using this data, \code{\link{both_male_long}} for the data in long format, and \code{\link{info}} for an overview of the actor exogenous attributes.
}
\keyword{dataset}
