% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{degree}
\alias{degree}
\title{degree}
\usage{
degree(edgelist, riskset, type)
}
\arguments{
\item{edgelist}{3-column edgelist (time, sender, receiver)}

\item{riskset}{2-column riskset (sender/actor 1, receiver/actor 2)}

\item{type}{1 = indegree_sender, 2 = indegree_receiver, 3 = 
outdegree_sender, 4 = outdegree_receiver, 5 = totaldegree_sender, 6 = 
totaldegree_receiver}
}
\value{
matrix (time x dyad)
}
\description{
A function to compute the indegree, outdegree and total degree effects.
}
\examples{
data(edgelistD)
out <- prepER(edgelistD, riskset = NULL, directed =  TRUE, type = FALSE)
el <- out$edgelist
rs <- out$riskset
indegree_sender <- degree(el, rs, type = 1)

}
