% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duremstats.R
\name{.parse_active_effects}
\alias{.parse_active_effects}
\title{Evaluate terms in a \code{duremstats} formula into effect config lists}
\usage{
.parse_active_effects(formula)
}
\arguments{
\item{formula}{Formula such as
\code{~ activeTie() + activeOutdegreeSender(scaling = "std")}}
}
\value{
A list of named lists, each with \code{$effect}, \code{$scaling},
  and \code{$consider_type}.
}
\description{
Evaluate terms in a \code{duremstats} formula into effect config lists
}
\keyword{internal}
