% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duremstats.R
\name{.validate_consider_type_durem}
\alias{.validate_consider_type_durem}
\title{Validate consider_type for active-state effects}
\usage{
.validate_consider_type_durem(consider_type)
}
\description{
Validate consider_type for active-state effects
}
\keyword{internal}
