% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dyadstat}
\alias{dyadstat}
\title{dyadstat}
\usage{
dyadstat(values, type, edgelist, riskset, equal_val)
}
\arguments{
\item{values}{3-column matrix (id, change time, value). Actor ids should 
correspond to actor ids in the edgelist and riskset. Change time is zero 
for all entries if the covariate is time invariate.}

\item{type}{1 = same, 2 = difference, 3 = mean, 4 = min, 5 = max, 
6 = both equal to}

\item{edgelist}{3-column edgelist (time, sender, receiver)}

\item{riskset}{2-column riskset (sender/actor 1, receiver/actor 2).}

\item{equal_val}{value}
}
\value{
matrix (time x dyad)
}
\description{
A function to transform exogenous actor covariate variables in dyad 
statistics.
}
\examples{
data(edgelistU)
data(covar)
out <- prepER(edgelist = edgelistU, directed = FALSE)
el <- out$edgelist
rs <- out$riskset
ac <- out$actors    
covar$id <- ac$id[match(covar$id, ac$name)]
covar <- as.matrix(covar)
stat <- dyadstat(values = covar[,c(1:3)], type = 1, edgelist = el, 
    riskset = rs, NA)

}
