% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{equate}
\alias{equate}
\title{equate}
\usage{
equate(variables, equal_val, covariates)
}
\arguments{
\item{variables}{character vector: names of one or more columns in the 
\code{covariates} object for which the statistic has to be computed.}

\item{equal_val}{numeric vector.}

\item{covariates}{an object of class \code{"\link[base]{data.frame}"} that 
contains the exogenous covariates. Each row should refer to one actor. The 
first column should contain the actor id (corresponding to the actor id's in 
the edgelist). The second column refers to the time when covariates change 
(set to zero if none of the covariates vary over time). Subsequent columns 
contain the covariates that are called in the specifications of exogenous 
statistics.}
}
\description{
Specifies the statistic for an `equate_to`-effect in the \code{formula} 
argument of \code{\link{remstats}}.
}
\details{
An `equate_to`-effect refers to an exogenous actor attribute that affects 
dyad \emph{(i,j)}'s rate of interacting based on whether both their values 
are equal to a specific value on this attribute. The statistic at timepoint 
\emph{t} for dyad \emph{(i,j)} is equal to one if the values of actor 
\emph{i} and \emph{j} on the attribute at timepoint \emph{t} are both equal 
to the specified value and equal to zero if not.
}
\examples{
data(history)
data(info)
remstats(~ equate(c("age", "age"), c(0,1), info), 
 edgelist = history)

}
