% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{history}
\alias{history}
\title{Simulated relational event history}
\format{
A dataframe with 112 rows and 5 variables:
\describe{
 \item{time}{time of the event since onset of observation (e.g., in minutes)}
 \item{actor1}{the first actor involved in the event}
 \item{actor2}{the second actor involved in the event}
 \item{setting}{the setting for the event}
 \item{weight}{the intensity of the event (e.g., based on the duration)}
}
}
\usage{
history
}
\description{
A dataset containing a small example of a relational event history. Data is 
simulated.
}
\seealso{
\code{\link{info}} for exogenous information on the actors in 
the social network.
}
\keyword{datasets}
