% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{indegreeSender}
\alias{indegreeSender}
\alias{degree}
\alias{indegree}
\title{indegreeSender}
\usage{
indegreeSender(scaling = c("as.is", "prop", "std"), consider_type = FALSE)
}
\arguments{
\item{scaling}{the method for scaling the degree statistic. Default is 
to not scale the statistic but keep the raw counts 'as.is'. Alternatively, 
scaling of the raw degree counts by the number of past events at time t can 
be requested with 'prop' or standardization of the raw degree counts per 
time point can be requested with 'std'.}

\item{consider_type}{logical, indicates whether to count the degrees 
separately for each event type (TRUE) or sum degrees across different event 
types (FALSE, default).}
}
\description{
Specifies the statistic for an `indegreeSender` effect in the \code{effects} 
argument of \code{\link{tomstats}} or the \code{rateEffects} argument of 
\code{\link{aomstats}}.
}
\details{
An indegree of the sender effect refers to the tendency for actors to send 
events if they have received more past events. The statistic at timepoint 
\emph{t} for dyad \emph{(i,j)} is equal to the number of events received by 
actor \emph{i} before timepoint \emph{t}. Note that the 'indegreeSender' 
effect is only defined for directed events. 

Optionally, a scaling method can be set with \code{scaling}. By scaling the 
degree count by the total number of past events, the statistic refers to the 
fraction of past events that were received by actor i. At the first time 
point, when no events did previously occur, it is assumed that every actor 
is equally likely to send a message and the statistic is set equal to 1/n, 
where n refers to the number of actors.
}
\examples{
effects <- ~ indegreeSender()
tomstats(effects, edgelist = history)
aomstats(rateEffects = effects, edgelist = history)

}
\seealso{
\code{\link{indegreeReceiver}}, \code{\link{outdegreeSender}}, 
\code{\link{outdegreeReceiver}}, \code{\link{totaldegreeSender}}, or
\code{\link{totaldegreeReceiver}} for other types of degree effects.
}
