% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{inertia_type}
\alias{inertia_type}
\title{inertia_type}
\usage{
inertia_type(edgelist, riskset, weights, standardize)
}
\arguments{
\item{edgelist}{4-column edgelist (time, sender, receiver, type)}

\item{riskset}{3-column riskset (sender/actor 1, receiver/actor 2, type)}

\item{weights}{vector (length edgelist)}

\item{standardize}{logical}
}
\value{
matrix (time x dyad)
}
\description{
A function to compute the inertia_type effect.
}
\examples{
data(edgelistDT)
out <- prepER(edgelist = edgelistDT, type = TRUE)
edgelist <- out$edgelist
riskset <- out$riskset
equal_weights <- rep(1, nrow(edgelist))
stat <- inertia_type(edgelist, riskset, equal_weights, standardize = FALSE)

}
