% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{inertia_typeMW}
\alias{inertia_typeMW}
\title{inertia_typeMW}
\usage{
inertia_typeMW(full_edgelist, window_edgelist, window_length, riskset,
  full_weights, standardize)
}
\arguments{
\item{full_edgelist}{4-column edgelist (time, sender, receiver, type).}

\item{window_edgelist}{4-column edgelist (time, sender, receiver, type).}

\item{window_length}{numeric value.}

\item{riskset}{3-column riskset (sender/actor 1, receiver/actor 2, type).}

\item{full_weights}{vector (length full_edgelist).}

\item{standardize}{logical.}
}
\value{
matrix (time x dyad)
}
\description{
A function to compute the inertia_type effect for a moving window REM.
}
\examples{
windows <- data.frame(start = seq(0, 900, 75), end = seq(100, 1000, 75))
data(edgelistUT)
window_edgelist <- edgelistUT[edgelistUT$time > windows$start[2] & 
    edgelistUT$time <= windows$end[2],]
out <- prepER(edgelist = edgelistUT, directed = FALSE, type = TRUE)
full_el <- out$edgelist
rs <- out$riskset
ac <- out$actors
out <- prepER(window_edgelist, directed = FALSE, type = TRUE, 
   actors = ac[,2])
window_el <- out$edgelist
equal_weights <- rep(1, nrow(full_el))
stat <- inertia_typeMW(full_edgelist = full_el, window_edgelist = 
    window_el, window_length = 100, riskset = rs, 
    full_weights = equal_weights, standardize = FALSE)

}
