% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{inertiaMW}
\alias{inertiaMW}
\title{inertiaMW}
\usage{
inertiaMW(full_evls, window_evls, window_length, riskset, full_weights,
  standardize)
}
\arguments{
\item{full_evls}{2-column edgelist (event, time) in relevent::rem format.}

\item{window_evls}{2-column edgelist (event, time) in relevent::rem format.}

\item{window_length}{numeric value.}

\item{riskset}{2-column riskset (sender/actor 1, receiver/actor 2).}

\item{full_weights}{vector (length full_evls).}

\item{standardize}{logical.}
}
\value{
matrix (time x dyad)
}
\description{
A function to compute the inertia effect.
}
\examples{
data(edgelistU)
windows <- data.frame(start = seq(0, 900, 75), end = seq(100, 1000, 75))
window_edgelist <- edgelistU[edgelistU$time > windows$start[2] & 
    edgelistU$time <= windows$end[2],]
out <- prepER(edgelist = edgelistU, directed = FALSE)
full_el <- out$edgelist
rs <- out$riskset
ac <- out$actors
out <- prepER(window_edgelist, directed = FALSE, actors = ac[,2])
window_el <- out$edgelist
full_evls <- prepEvls(full_el, rs)
window_evls <- prepEvls(window_el, rs)
stat <- inertiaMW(full_evls = full_evls, window_evls = window_evls, 
window_length = 100, riskset = rs, full_weights = rep(1, nrow(el)), 
standardize = FALSE)

}
