% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remstats_durem.R
\name{is.remstats_durem}
\alias{is.remstats_durem}
\title{Test whether an object is a \code{remstats_durem}}
\usage{
is.remstats_durem(x)
}
\arguments{
\item{x}{Any R object.}
}
\value{
\code{TRUE} if \code{x} inherits from \code{"remstats_durem"}.
}
\description{
Test whether an object is a \code{remstats_durem}
}
