% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{itp}
\alias{itp}
\title{itp}
\usage{
itp(scaling = c("as.is", "std"), consider_type = FALSE)
}
\arguments{
\item{scaling}{the method for scaling the triad statistic. Default is to not
scale the statistic but keep the raw counts 'as.is'. Alternatively,
standardization of the raw counts per time point can be requested
with 'std'.}

\item{consider_type}{logical, indicates whether to count the two-paths
separately for each event type (TRUE) or sum across different event
types (FALSE, default).}
}
\description{
Specifies the statistic for an incoming two-path effect in the tie-oriented
model or the receiver choice step of the actor-oriented model.
}
\details{
An incomping two-path effect refers to the tendency of dyads to interact if
they have more past incoming two-paths between them. The statistic at
timepoint \emph{t} for dyad \emph{(i,j)} is equal to the minimum of past
\emph{(j,h)}, \emph{(h,i)} events, summed over all actors \emph{h}.
Optionally, a scaling method can be set with \code{scaling}. Note that an
'itp' effect is only defined for directed events.
}
\examples{
library(remify)
rehObject <- remify(history)
effects <- ~ itp()
remstats(reh = rehObject, tie_effects = effects)
remstats(reh = rehObject, receiver_effects = effects)
tomstats(effects, reh = rehObject)
aomstats(receiver_effects = effects, reh = rehObject)

}
\seealso{
\code{\link{otp}}, \code{\link{osp}}, or \code{\link{isp}} for
other types of triadic effects for directed relational events and
\code{\link{sp}} or \code{\link{spUnique}} for triadic effects for
undirected relational events.
}
