% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{osp}
\alias{osp}
\title{osp}
\usage{
osp(
  scaling = c("counts", "standardize"),
  memory_value = Inf,
  with_type = FALSE,
  event_weights = NULL
)
}
\arguments{
\item{scaling}{the method for scaling the triad statistic. Options are one 
of \code{"counts"} (default option, gives the raw triad counts at time t), 
or \code{"standardize"} (in which raw triad counts are standardized per time 
point).}

\item{memory_value}{numeric value. Specifies the time after which events are 
no longer included in the degree count (default: all past events are 
considered). Note: make sure memory_value is specified in the same time unit 
as the time for the events in the edgelist.}

\item{with_type}{logical value. If TRUE, past events of different types c 
are counted separately for every (i,j,c) event in the risk set. If FALSE 
(default), the type of the events is not considered in the degree counts. 
Note: with_type can be FALSE even when a distinction between different event 
types is made in the risk set.}

\item{event_weights}{vector with numeric values that indicate the intensity 
of the events in the edgelist. If event weights are supplied, the degree 
statistic is equal to the sum of the intensity of past events instead of the 
number of past events.}
}
\description{
Specifies the statistic for an outgoing shared partners effect in the 
\code{formula} argument of \code{\link{remstats}}.
}
\details{
An outgoing shared partners effect refers to the tendency of dyads to 
interact if they have more past outgoing shared partners between them. The 
statistic at timepoint \emph{t} for dyad \emph{(i,j)} is equal to the 
minimum of past \emph{(i,h)}, \emph{(j,h)} events, summed over all actors 
\emph{h}. Optionally, a scaling method can be set with \code{scaling}, 
events that happened a certain number of time units ago can be disregarded 
in the count by setting \code{memory_value}, events of different types can 
be counted separately by setting \code{with_type = TRUE} or past events can 
be weighted by setting \code{event_weights}.
}
\examples{
data(history)
remstats(~ osp(), edgelist = history)

}
\seealso{
\code{\link{otp}}, \code{\link{itp}}, or \code{\link{isp}} for 
other types of triadic effects for directed relational events and 
\code{\link{sp}} or \code{\link{spUnique}} for triadic effects for 
undirected relational events.
}
