% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{outdegreeReceiver}
\alias{outdegreeReceiver}
\title{outdegreeReceiver}
\usage{
outdegreeReceiver(
  scaling = c("counts", "total", "standardize"),
  memory_value = Inf,
  with_type = FALSE,
  event_weights = NULL
)
}
\arguments{
\item{scaling}{the method for scaling the degree statistic. Options are one 
of \code{"counts"} (default option, gives the raw degree counts at time t), 
\code{"total"} (in which raw degree counts are divided by the number of past 
events at time t), and \code{"standardize"} (in which raw degree counts are 
standardized per time point).}

\item{memory_value}{numeric value. Specifies the time after which events are 
no longer included in the degree count (default: all past events are 
considered). Note: make sure memory_value is specified in the same time unit 
as the time for the events in the edgelist.}

\item{with_type}{logical value. If TRUE, past events of different types c 
are counted separately for every (i,j,c) event in the risk set. If FALSE 
(default), the type of the events is not considered in the degree counts. 
Note: with_type can be FALSE even when a distinction between different event 
types is made in the risk set.}

\item{event_weights}{vector with numeric values that indicate the intensity 
of the events in the edgelist. If event weights are supplied, the degree 
statistic is equal to the sum of the intensity of past events instead of the 
number of past events.}
}
\description{
Specifies the statistic for an outdegree of the receiver effect in the 
\code{formula} argument of \code{\link{remstats}}.
}
\details{
An outdegree of the sender effect refers to the tendency for actors to 
receive events if they have send more past events. The statistic at 
timepoint \emph{t} for dyad \emph{(i,j)} is equal to the number of events 
send by actor \emph{j} before timepoint \emph{t}. Optionally, a scaling 
method can be set with \code{scaling}, events that happened a certain number 
of time units ago can be disregarded in the count by setting 
\code{memory_value}, events of different types can be counted separately by 
setting \code{with_type = TRUE} or past events can be weighted by setting 
\code{event_weights}.
}
\examples{
data(history)
remstats(~ outdegreeReceiver(), edgelist = history)

}
\seealso{
\code{\link{indegreeSender}}, \code{\link{indegreeReceiver}}, 
\code{\link{outdegreeSender}}, \code{\link{totaldegreeSender}}, or
\code{\link{totaldegreeReceiver}} for other types of degree effects.
}
