% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{overlap}
\alias{overlap}
\title{overlap}
\usage{
overlap(
  consider_type = TRUE,
  scaling = c("none", "std", "std_abs", "std_unique")
)
}
\arguments{
\item{consider_type}{logical; whether to compute per type.}

\item{scaling}{character; scaling method.}
}
\description{
Overlap effect: proportion of shared interaction partners.
Symmetric: overlap(A,B) == overlap(B,A), also for directed networks.
}
\details{
The value is 0 when at least one of the actors in the dyad has not interacted 
with anyone yet (because there are no shared partners at all).
}
