% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepEvls.R
\name{prepEvls}
\alias{prepEvls}
\title{prepEvls}
\usage{
prepEvls(edgelist, riskset, type)
}
\arguments{
\item{edgelist}{[matrix] or [dataframe], should minimally contain the time, 
sender/actor 1 and receiver/actor 2 in the first three columns respectively. 
If the riskset contains typed relational events, the fourth column should 
contain the event type.}

\item{riskset}{[matrix] or [dataframe], should minimally contain sender/
actor 1 and receiver/actor 2 in the first two columns, respectively. If it 
contains typed relational events, the third column should contain the event 
type.}

\item{type}{[logical] do relational events in the riskset consider an 
action type (type = TRUE) or not (type = FALSE, default).}
}
\value{
evls [matrix] edgelist transformed in the format that is required 
for estimation by relevent::rem() with in the first column the relational 
event ID and in the second column the time.
}
\description{
A helper function that matches events in the edgelist with events in the 
riskset to create an evls, i.e., an edgelist in the format required by 
relevent::rem.
}
\examples{
data(edgelistD)
out <- prepER(edgelistD, riskset = NULL, directed = TRUE, type = FALSE)
el <- out$edgelist
rs <- out$riskset
evls <- prepEvls(el, rs, type = FALSE)

}
