% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remstats_durem.R
\name{print.remstats_durem}
\alias{print.remstats_durem}
\title{Print method for \code{remstats_durem}}
\usage{
\method{print}{remstats_durem}(x, ...)
}
\arguments{
\item{x}{A \code{remstats_durem} object.}

\item{...}{Ignored.}
}
\description{
Print method for \code{remstats_durem}
}
