% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{psABBY}
\alias{psABBY}
\title{psABBY}
\usage{
psABBY(with_type = FALSE)
}
\arguments{
\item{with_type}{logical value. If TRUE, the pshift is set only for those
dyads in the riskset with the same event type as the last event. If FALSE 
(default), pshifts are set regardless of the event types of the dyads in 
the riskset.}
}
\description{
Specifies the statistic for a pshift AB-BY effect in the \code{formula} 
argument of \code{\link{remstats}}.
}
\details{
The AB-BY pshift effect refers to one of Gibson's (2003) dyadic 
participation shifts. The AB-BY pshift refers to a tendency for turn 
receiving (here, the next sender is the current receiver and the next 
receiver is not in the current event). For each timepoint t, the psABBY 
statistic is equal to one for the dyads that will create the participation 
shift if they would occur in the edgelist at time t and equal to zero for 
the dyads that will not create this participation shift.
}
\examples{
data(history)
remstats(~ psABBY(), edgelist = history)

}
\seealso{
\code{\link{psABBA}}, \code{\link{psABXA}}, \code{\link{psABXB}}, 
\code{\link{psABXY}} or \code{\link{psABAY}} for other dyadic participation 
shifts.
}
