% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{psABXY}
\alias{psABXY}
\title{psABXY}
\usage{
psABXY(with_type = FALSE)
}
\arguments{
\item{with_type}{logical value. If TRUE, the pshift is set only for those
dyads in the riskset with the same event type as the last event. If FALSE 
(default), pshifts are set regardless of the event types of the dyads in 
the riskset.}
}
\description{
Specifies the statistic for a pshift AB-XY effect in the \code{formula} 
argument of \code{\link{remstats}}.
}
\details{
The AB-XY pshift effect refers to one of Gibson's (2003) dyadic 
participation shifts. The AB-XY pshift refers to a tendency for turn 
usurping (here, the next sender and the next receiver are not in the current 
event). For each timepoint t, the psABXY statistic is equal to one for the 
dyads that will create the participation shift if they would occur in the 
edgelist at time t and equal to zero for the dyads that will not create this 
participation shift.
}
\examples{
data(history)
remstats(~ psABXY(), edgelist = history)

}
\seealso{
\code{\link{psABBA}}, \code{\link{psABBY}}, \code{\link{psABXA}}, 
\code{\link{psABXB}} or \code{\link{psABAY}} for other dyadic participation 
shifts.
}
