% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{psABYAB}
\alias{psABYAB}
\title{psABYAB}
\usage{
psABYAB(consider_type = TRUE)
}
\arguments{
\item{consider_type}{logical, indicates whether to consider the event type
in determining which dyads create a pshift (TRUE, default) or not (FALSE).}
}
\value{
List with all information required by `remstats::remstats()` to
compute the statistic.
}
\description{
Specifies the statistic for a composite pshift ABY-AB effect.
}
\details{
Composite participation shift defined as \code{psABAY()} + \code{psABXB()}.
The supplied \code{consider_type} argument is used for both components.
}
\examples{
reh <- remify::remify(history, model = "tie")
effects <- ~ psABYAB()
remstats(reh = reh, tie_effects = effects)

}
