% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pshift}
\alias{pshift}
\title{pshift}
\usage{
pshift(edgelist, riskset, type)
}
\arguments{
\item{edgelist}{3-column edgelist (time, sender, receiver)}

\item{riskset}{2-column riskset (sender/actor 1, receiver/actor 2)}

\item{type}{(1 = AB-BA, 2 = AB-BY, 3 = AB-XA, 4 = AB-XB, 5 = AB-XY, 
6 = AB-AY)}
}
\value{
matrix (time x dyad)
}
\description{
A function to compute Gibson's (2003) dyadic participation shifts.
}
\examples{
data(edgelistD)
out <- prepER(edgelistD)
ABBA <- pshift(edgelist = out$edgelist, riskset = out$riskest, type = 1)

}
