% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{recency}
\alias{recency}
\title{recency}
\usage{
recency(edgelist, actors, type)
}
\arguments{
\item{edgelist}{3-column edgelist (time, sender, receiver)}

\item{actors}{vector of integers indicating the identity of the actors in the network}

\item{type}{equals 1 for outgoing and 2 for incoming recency effect}
}
\value{
matrix (time x dyad)
}
\description{
A function to compute the rank-based recency effect, as in section 2.2.5 of Butts (2008).
}
\examples{
data(edgelistD)
out <- prepER(edgelist = edgelistD)
edgelist <- out$edgelist
actors <- out$actors$id
type <- 1
stat <- recency(edgelist, actors, type)

}
