% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{recencySender}
\alias{recencySender}
\title{recencySender}
\usage{
recencySender(memory_value = Inf)
}
\arguments{
\item{memory_value}{numeric value. Specifies the time after which events are 
no longer included in the statistic count (default: all past events are 
considered). Note: make sure memory_value is specified in the same time unit 
as the time for the events in the edgelist.}
}
\description{
Specifies the statistic for a recency sender effect in the 
\code{formula} argument of \code{\link{remstats}}.
}
\details{
The recencySender effect refers to a recency statistic similar to Vu et 
al. (2017) and Mulder and Leenders (2019). For each timepoint t, for 
directed dyad (i,j) the statistic is equal to 1/(the time that has past 
since sender i was last active + 1).
}
\examples{
data(history)
remstats(~ recencySender(), edgelist = history)

}
\seealso{
\code{\link{recencyReceiver}} and \code{\link{recencyContinue}}
}
