% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{recencySendReceiver}
\alias{recencySendReceiver}
\title{recencySendReceiver}
\usage{
recencySendReceiver(consider_type = FALSE)
}
\arguments{
\item{consider_type}{logical, indicates whether to compute the recency
separately for each event type (TRUE) or regardless of event types (FALSE,
default).}
}
\description{
Specifies the statistic for a recency send of receiver effect in the
\code{effects} argument of \code{\link{tomstats}} or the
\code{receiver_effects} argument of \code{\link{aomstats}}.
}
\details{
The recencySendReceiver effect refers to a recency statistic similar to what
is described in Vu et al. (2017) and Mulder and Leenders (2019). For each
timepoint t, for directed dyad (i,j) the statistic is equal to 1/(the time
that has past since receiver j was last active as sender + 1). Note that the
'recencySendReceiver' effect is only defined for directed events.
}
\examples{
effects <- ~ recencySendReceiver()
reh_tie <- remify::remify(history, model = "tie")
remstats(tie_effects = effects, reh = reh_tie)

reh_actor <- remify::remify(history, model = "actor")
remstats(receiver_effects = effects, reh = reh_actor)

}
\seealso{
\code{\link{rrankSend}}, \code{\link{rrankReceive}},
\code{\link{recencySendSender}}, \code{\link{recencyReceiveSender}},
\code{\link{recencyReceiveReceiver}} and \code{\link{recencyContinue}} for
other type of recency effects
}
