% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{recencySendSender}
\alias{recencySendSender}
\title{recencySendSender}
\usage{
recencySendSender(consider_type = FALSE, param = NULL)
}
\arguments{
\item{consider_type}{logical, indicates whether to compute the recency
separately for each event type (TRUE) or regardless of event types (FALSE, 
default).}

\item{param}{numeric value or function with time parameter to be used for 
generation of relational event histories with \code{\link{remulate}}.}
}
\description{
Specifies the statistic for a recency send of sender effect in the 
\code{effects} argument of \code{\link{tomstats}} or the \code{sender_effects} 
argument of \code{\link{aomstats}}.
}
\details{
The recencySendSender effect refers to a recency statistic similar to what 
is described in Vu et al. (2017) and Mulder and Leenders (2019). For each 
timepoint t, for directed dyad (i,j) the statistic is equal to 1/(the time 
that has past since sender i was last active as sender + 1). Note that the 
'recencySendSender' effect is only defined for directed events.
}
\examples{
effects <- ~ recencySendSender()
tomstats(effects, edgelist = history)
aomstats(sender_effects = effects, edgelist = history)

}
\seealso{
\code{\link{rrankSend}}, \code{\link{rrankReceive}}, 
\code{\link{recencySendReceiver}}, \code{\link{recencyReceiveSender}}, 
\code{\link{recencyReceiveReceiver}} and \code{\link{recencyContinue}} for 
other type of recency effects
}
