% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{recenyReceiver}
\alias{recenyReceiver}
\title{recenyReceiver}
\usage{
recenyReceiver(memory_value = NULL)
}
\arguments{
\item{memory_value}{numeric value. Specifies the time after which events are 
no longer included in the statistic count (default: all past events are 
considered). Note: make sure memory_value is specified in the same time unit 
as the time for the events in the edgelist. 

seealso

examples}
}
\description{
Description
}
\details{
Details
}
