% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{reciprocity}
\alias{reciprocity}
\title{reciprocity}
\usage{
reciprocity(edgelist, riskset)
}
\arguments{
\item{edgelist}{3-column edgelist (time, sender, receiver)}

\item{riskset}{2-column riskset (sender/actor 1, receiver/actor 2)}
}
\value{
matrix (time x dyad)
}
\description{
A function to compute the indegree, outdegree and total degree effects.
}
\examples{
data(edgelistD)
out <- prepER(edgelist = edgelistD, directed = TRUE, type = FALSE, 
    riskset = NULL, actors = NULL)
el <- out$edgelist
rs <- out$riskset
stat <- reciprocity(el, rs)

}
