\name{remstats-package}
\alias{remstats-package}
\alias{remstats}
\docType{package}
\title{
\packageTitle{remstats}
}
\description{
\packageDescription{remstats}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{remstats}
\packageIndices{remstats}
An overview of how to use the package, including the most important functions
}
\author{
\packageAuthor{remstats}

Maintainer: \packageMaintainer{remstats}
}
\references{
Literature or other references for background information
}
\keyword{ package }
\seealso{
Optional links to other man pages
}
