% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{remStatsC}
\alias{remStatsC}
\title{remStatsC}
\usage{
remStatsC(effects, edgelist, riskset, evls, actors, sender_values,
  receiver_values, weights)
}
\description{
A function to compute statistics and combine the statistics in an array
prepared for estimation of a REM with relevent::rem(). Used internally 
in remStats.
}
\details{
param:
[effects] integer vector (effects)
[edgelist] 3-column edgelist (time, sender, receiver)
[riskset] 2-column riskset (sender/actor 1, receiver/actor 2)
[evls] 2-column edgelist (event, time) in relevent::rem format
[actors] vector with numeric actor IDs (correspod to edgelist, riskset)
[sender_values] matrix (id, time, sender covariate values)
[receiver_values] matrix(id, time, receiver covariate values)
[weights] vector (length evls)

return:
[statistics] 3-dimensional array (event time x risk set entry x statistic)
}
