% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{remstatsCpp}
\alias{remstatsCpp}
\title{remstatsCpp}
\usage{
remstatsCpp(effects, standardize, edgelist, riskset, evls, actors,
  covariates, event_effect, weights, equal_val, int_positions)
}
\description{
A function to compute statistics and combine the statistics in an array
prepared for estimation of a REM with relevent::rem(). Used internally 
in remstats.
}
\details{
param:
[effects] integer vector (effects)
[standardize] logical, indicates whether statistics for endogenous effects 
should be standardized
[edgelist] 3-column edgelist (time, sender, receiver)
[riskset] 2-column riskset (sender/actor 1, receiver/actor 2)
[evls] 2-column edgelist (event, time) in relevent::rem format
[actors] vector with numeric actor IDs (correspod to edgelist, riskset)
[covariates] List with matrices
    0: [sender_values] matrix (id, time, covariate values)
    1: [receiver_values] matrix(id, time, covariate values)
    2: [same] matrix(id, time, covariate values)
    3: [difference] matrix(id, time, covariate values)
    4: [mean] matrix(id, time, covariate values)
    5: [min] matrix(id, time, covariate values)
    6: [max] matrix(id, time, covariate values)
    7: [both_equal_to] matrix(id, time, covariate values)
[event_effect] matrix (event effect per column)
[weights] vector (length evls) 
[equal_val] vector (length ncol both_equal_to minus 2)
[int_positions] matrix (effect 1, effect 2)

return:
[statistics] 3-dimensional array (event time x risk set entry x statistic)
}
