% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{remstatsMWCpp}
\alias{remstatsMWCpp}
\title{remstatsMWCpp}
\usage{
remstatsMWCpp(
  effects,
  standardize,
  full_edgelist,
  window_edgelist,
  window_length,
  riskset,
  actors,
  covariates,
  event_effect,
  types,
  full_weights,
  equal_val,
  int_positions
)
}
\arguments{
\item{effects}{integer vector (effects)}

\item{standardize}{logical, indicates whether statistics for endogenous 
effects should be standardized}

\item{full_edgelist}{3-column edgelist (time, sender, receiver)}

\item{window_edgelist}{3-column edgelist (time, sender, receiver)}

\item{window_length}{numeric value.}

\item{riskset}{2-column riskset (sender/actor 1, receiver/actor 2)}

\item{actors}{vector with numeric actor IDs (correspod to edgelist, riskset)}

\item{covariates}{List with matrices
0: [sender_values] matrix (id, time, covariate values)
1: [receiver_values] matrix(id, time, covariate values)
2: [same] matrix(id, time, covariate values)
3: [difference] matrix(id, time, covariate values)
4: [mean] matrix(id, time, covariate values)
5: [min] matrix(id, time, covariate values)
6: [max] matrix(id, time, covariate values)
7: [both_equal_to] matrix(id, time, covariate values)}

\item{event_effect}{matrix (event effect per column)}

\item{types}{vector}

\item{full_weights}{vector (length evls)}

\item{equal_val}{vector (length ncol both_equal_to minus 2)}

\item{int_positions}{matrix (effect 1, effect 2)}
}
\value{
statistics 3-dimensional array (event time x risk set entry x 
statistic)
}
\description{
A function to compute statistics and combine the statistics in an array
prepared for estimation of a moving window REM with relevent::rem(). Used 
internally in remstatsMW.
}
