% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{send}
\alias{send}
\title{send}
\usage{
send(variables, covariates)
}
\arguments{
\item{variables}{character vector: names of one or more columns in the 
\code{covariates} object for which the statistic has to be computed.}

\item{covariates}{an object of class \code{"\link[base]{data.frame}"} that 
contains the exogenous covariates. Each row should refer to one actor. The 
first column should contain the actor id (corresponding to the actor id's in 
the edgelist). The second column refers to the time when covariates change 
(set to zero if none of the covariates vary over time). Subsequent columns 
contain the covariates that are called in the specifications of exogenous 
statistics.}
}
\description{
Specifies the statistic for a send effect in the \code{formula} argument 
of \code{\link{remstats}}.
}
\details{
A send effect refers to an exogenous actor attribute that affects actor 
i's rate of sending events. The statistic at timepoint \emph{t} is equal to 
the value of the exogenous attribute for actor \emph{i} at time \emph{t} for 
all dyads in the riskset that have actor \emph{i} as sender. A send effect 
is only defined for directed relational events.
}
\examples{
data(history)
data(info)
remstats(~ send(c("extraversion", "agreeableness"), info), 
 edgelist = history)

}
