% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{spUnique}
\alias{spUnique}
\title{spUnique}
\usage{
spUnique(
  scaling = c("counts", "standardize"),
  memory_value = Inf,
  with_type = FALSE
)
}
\arguments{
\item{scaling}{the method for scaling the triad statistic. Options are one 
of \code{"counts"} (default option, gives the raw triad counts at time t), 
or \code{"standardize"} (in which raw triad counts are standardized per time 
point).}

\item{memory_value}{numeric value. Specifies the time after which events are 
no longer included in the degree count (default: all past events are 
considered). Note: make sure memory_value is specified in the same time unit 
as the time for the events in the edgelist.}

\item{with_type}{logical value. If TRUE, past events of different types c 
are counted separately for every (i,j,c) event in the risk set. If FALSE 
(default), the type of the events is not considered in the degree counts. 
Note: with_type can be FALSE even when a distinction between different event 
types is made in the risk set.}
}
\description{
Specifies the statistic for a unique shared partners effect in the 
\code{formula} argument of \code{\link{remstats}}.
}
\details{
A unique shared partners effect refers to the tendency of dyads to interact 
if they have more past unique shared partners between them. The statistic is 
defined for undirected relational events. The statistic at timepoint 
\emph{t} for dyad \emph{(i,j)} is equal to the number of unique actors 
\emph{h} that both actors \emph{i} and \emph{j} interacted with in the past. 
Optionally, a scaling method can be set with \code{scaling}, events that 
happened a certain number of time units ago can be disregarded in the count 
by setting \code{memory_value} or events of different types can be counted 
separately by setting \code{with_type = TRUE}.
}
\examples{
data(history)
remstats(~ spUnique(), edgelist = history, directed = FALSE)

}
\seealso{
\code{\link{sp}} for another type of triadic effect for 
undirected relational events and \code{\link{otp}}, \code{\link{itp}}, 
\code{\link{osp}}, or \code{\link{isp}} for triadic effects for directed 
relational events.
}
