% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remstats_durem.R
\name{summary.remstats_durem}
\alias{summary.remstats_durem}
\title{Summary method for \code{remstats_durem}}
\usage{
\method{summary}{remstats_durem}(object, ...)
}
\arguments{
\item{object}{A \code{remstats_durem} object.}

\item{...}{Ignored.}
}
\description{
Summary method for \code{remstats_durem}
}
