% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{tempEquiv}
\alias{tempEquiv}
\title{tempEquiv}
\usage{
tempEquiv(
  consider_type = TRUE,
  approach = c("l1", "pearson"),
  scaling = c("none", "std", "std_abs", "std_unique")
)
}
\arguments{
\item{consider_type}{logical; whether to compute per type.}

\item{approach}{either "l1" (usually best) or "pearson"}

\item{scaling}{character; scaling method.}
}
\description{
Temporal structural equivalence effect
}
\details{
Temporal structural equivalence: for any point in time, this effect measures 
the extent to which actors in a dyad are equivalent in their interaction history.

This takes the vector of inertia for the two actors in the dyad (within the 
riskset) and either calculates the L1 distance or the pearson correlation 
between them. The L1 distance is transformed to an equivalence as 1 - L1/m, 
with m the length of the vector. 
The L1 measure is calculated on the inertias after they have been divided by 
the highest inertia across all dyads at that time point. As a result, the 
measure is within the unity interval, with 0 being perfect dissimilarity and 
1 for perfect temporal structural equivalence.

Note that the pearson correlation is not always defined.
}
