% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{triad}
\alias{triad}
\title{triad}
\usage{
triad(actors, edgelist, riskset, type, standardize)
}
\arguments{
\item{actors}{vector with numeric actor IDs (correspod to edgelist, riskset)}

\item{edgelist}{3-column edgelist (time, sender, receiver)}

\item{riskset}{2-column riskset (sender/actor 1, receiver/actor 2)}

\item{type}{(1 = outgoing two-path, 2 = incoming two-path, 3 = outbound 
shared partners, 4 = inbound shared partners)}

\item{standardize}{logical}
}
\value{
matrix (time x dyad)
}
\description{
A function to compute the triadic effects.
}
\examples{
data(edgelistD)
out <- prepER(edgelist = edgelistD, directed = TRUE, type = FALSE, 
    riskset = NULL, actors = NULL)
el <- out$edgelist
rs <- out$riskset
ac <- out$actors[,1]
otp <- triad(ac, el, rs, type = 1, standardize = FALSE)

}
