% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{triadU}
\alias{triadU}
\title{triadU}
\usage{
triadU(actors, edgelist, riskset)
}
\arguments{
\item{actors}{vector with numeric actor IDs (correspod to edgelist, riskset)}

\item{edgelist}{3-column edgelist (time, sender, receiver)}

\item{riskset}{2-column riskset (sender/actor 1, receiver/actor 2)}
}
\value{
matrix (time x dyad)
}
\description{
A function to compute the shared partners effect for undirected relational 
events.
}
\examples{
data(edgelistU)
out <- prepER(edgelist = edgelistU, directed = FALSE, type = FALSE, 
    riskset = NULL, actors = NULL)
el <- out$edgelist
rs <- out$riskset
ac <- sort(unique(c(rs[,1], rs[,2])))
otp <- triadU(ac, el, rs)

}
