% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{triadU_type}
\alias{triadU_type}
\title{triadU_type}
\usage{
triadU_type(actors, edgelist, riskset, unique_sp, standardize)
}
\arguments{
\item{actors}{vector with numeric actor IDs (correspond to edgelist,
riskset)}

\item{edgelist}{4-column edgelist (time, sender, receiver, type)}

\item{riskset}{3-column riskset (sender/actor 1, receiver/actor 2, type)}

\item{unique_sp}{logical value}

\item{standardize}{logical value}
}
\value{
matrix (time x dyad)
}
\description{
A function to compute the (unique) shared partners effect for undirected 
relational events with types.
}
\examples{
data(edgelistUT)
out <- prepER(edgelist = edgelistUT, directed = FALSE, type = TRUE)
el <- out$edgelist
rs <- out$riskset
ac <- out$actors[,1]
stat <- triadU_type(ac, el, rs, unique_sp = FALSE, standardize = FALSE)

}
