% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{triadUMW}
\alias{triadUMW}
\title{triadUMW}
\usage{
triadUMW(actors, full_edgelist, window_edgelist, window_length, riskset,
  unique_sp, standardize)
}
\arguments{
\item{actors}{vector with numeric actor IDs (correspond to edgelist,
riskset)}

\item{full_edgelist}{3-column edgelist (time, sender, receiver)}

\item{window_edgelist}{3-column edgelist (time, sender, receiver)}

\item{window_length}{numeric value.}

\item{riskset}{2-column riskset (sender/actor 1, receiver/actor 2)}

\item{unique_sp}{logical value}

\item{standardize}{logical value}
}
\value{
matrix (time x dyad)
}
\description{
A function to compute the (unique) shared partners effect for undirected 
relational events when fitting a moving window REM.
}
\examples{
data(edgelistU)
windows <- data.frame(start = seq(0, 900, 75), end = seq(100, 1000, 75))
window_edgelist <- edgelistU[edgelistU$time > windows$start[2] & 
    edgelistU$time <= windows$end[2],]
out <- prepER(edgelist = edgelistU, directed = FALSE)
full_el <- out$edgelist
rs <- out$riskset
ac <- out$actors
out <- prepER(window_edgelist, directed = FALSE, actors = ac[,2])
window_el <- out$edgelist
stat <- triadUMW(actors = ac[,1], full_edgelist = full_el, 
    window_edgelist = window_el, window_length = 100, riskset = rs, 
    unique_sp = FALSE, standardize = FALSE)

}
