% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{typestat}
\alias{typestat}
\title{typestat}
\usage{
typestat(edgelist, riskset, type)
}
\arguments{
\item{edgelist}{4-column edgelist (time, sender, receiver, type)}

\item{riskset}{3-column riskset (sender/actor 1, receiver/actor 2, type)}

\item{type}{numeric value}
}
\value{
matrix (time x dyad)
}
\description{
A function to compute a type effect/type dummy.
}
\examples{
data(edgelistDT)
out <- prepER(edgelist = edgelistDT, type = TRUE)
edgelist <- out$edgelist
riskset <- out$riskset
types <- unique(edgelist[,4])
stat <- typestat(edgelist, riskset, types[1])

}
