% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{average}
\alias{average}
\title{average}
\usage{
average(variable, attributes = NULL, scaling = c("as.is", "std"))
}
\arguments{
\item{variable}{string with the name of the column in the
\code{attributes} object for which the statistic has to be computed.}

\item{attributes}{optionally, an object of class
\code{\link[base]{data.frame}} that contains the attribute, see
'Details.'}

\item{scaling}{the method for scaling the statistic. Default is to not scale
the statistic but keep it "as.is". Alternatively, standardization of the
statistic per time point can be requested with "std".}
}
\description{
Specifies the statistic for an "average" effect in the tie-oriented model or
the receiver choice step of the actor-oriented model. An "average" effect
refers to an exogenous actor attribute that affects dyad \emph{(i,j)}'s rate
of interacting (tie-oriented model) or actor \emph{j}'s probability of being
chosen as a receiver for the event send by the active sender \emph{i} at
time \emph{t} (actor-oriented model) based on the average of the values of
actors \emph{i} and \emph{j} on this attribute.
}
\details{
The statistic at timepoint \emph{t} for dyad \emph{(i,j)} is equal to the
average of the values of actor \emph{i} and \emph{j} on the attribute at
timepoint \emph{t}.

The \code{attributes} object should be constructed as a
\code{\link[base]{data.frame}} where each row refers to the attribute value
of actor \emph{i} at timepoint \emph{t}:
\itemize{
 \item{\code{id}: the actors' id}
 \item{\code{time}: the time when the attribute values change (set to a
column with only zero's if the attribute does not vary over time).}
 \item{}{the third column contains the attribute that is called in the
specification of the "average" effect (i.e., the column name should
correspond to the string that is supplied to the \code{variable} argument)}
}
Note that it is possible to omit the attributes object in the call of
\code{average()} and, instead, supply it in the call of \code{remstats()} for
multiple exogenous effects.
}
\examples{
library(remify)
rehObject <- reh(history)
effects <- ~ average("extraversion")
remstats(reh = rehObject, tie_effects = effects, attributes = info)
remstats(reh = rehObject, receiver_effects = effects, attributes = info)

}
