% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{ccp}
\alias{ccp}
\title{ccp}
\usage{
ccp(duration)
}
\arguments{
\item{duration}{vector with event durations.}
}
\description{
Statistic for a 'current common partner' effect.
}
\details{
The 'ccp' or 'current common partner' effect refers to the presence of a 
shared interaction between the actors in dyad (i, j) at time 't' with common 
other actors. This statistic can be calculated for undirected events with a 
duration in the tie-oriented model.
}
