% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{ccp}
\alias{ccp}
\title{ccp}
\usage{
ccp(duration)
}
\arguments{
\item{duration}{vector with event durations.}
}
\description{
Specifies the statistic for a 'common current partner' effect for the
tie-oriented model. NOTE: currently only defined for undirected, dyadic
events.
}
