% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{difference}
\alias{difference}
\title{difference}
\usage{
difference(variables, covariates)
}
\arguments{
\item{variables}{character vector: names of one or more columns in the 
\code{covariates} object for which the statistic has to be computed.}

\item{covariates}{an object of class \code{"\link[base]{data.frame}"} that 
contains the exogenous covariates. Each row should refer to one actor. The 
first column should contain the actor id (corresponding to the actor id's in 
the edgelist). The second column refers to the time when covariates change 
(set to zero if none of the covariates vary over time). Subsequent columns 
contain the covariates that are called in the specifications of exogenous 
statistics.}
}
\description{
Specifies the statistic for a difference effect in the \code{formula} 
argument of \code{\link{remstats}}.
}
\details{
A difference effect refers to an exogenous actor attribute that affects dyad 
\emph{(i,j)}'s rate of interacting based on the difference between their 
values on this attribute. The statistic at timepoint \emph{t} for dyad 
\emph{(i,j)} is equal to the absolute difference between the values of actor 
\emph{i} and \emph{j} on the attribute at timepoint \emph{t}.
}
\examples{
data(history)
data(info)
remstats(~ difference(c("extraversion", "agreeableness"), info), 
 edgelist = history)

}
