% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{event}
\alias{event}
\title{event}
\usage{
event(x, variableName = NULL, param = NULL)
}
\arguments{
\item{x}{vector with the event attribute}

\item{variableName}{optionally, a string indicating the variable name, used 
for the dimnames of the output statistics object}

\item{param}{numeric value or function with time parameter to be used for 
generation of relational event histories with \code{\link{remulate}}.}
}
\description{
Specifies the statistic for an `event` effect in the tie-oriented model.
}
\details{
An `event` effect refers to an exogenous event attribute that affects the 
waiting time between events. The statistic at timepoint \emph{t} is for all 
dyads in the riskset equal to the attribute of the event at timepoint 
\emph{t}.
}
\examples{
library(remify)
rehObject <- reh(history)
work <- ifelse(history$setting == "work", 1, 0)
effects <- ~ event(work, "setting.is.work")
remstats(edgelist = rehObject, tie_effects = effects, attributes = info)
tomstats(effects, edgelist = rehObject)

}
\seealso{
\code{\link{FEtype}}
}
