% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{event}
\alias{event}
\title{event}
\usage{
event(variable)
}
\arguments{
\item{variable}{string with the name of the column in the 
\code{edgelist} object supplied to \code{\link{remstats}} or 
\code{\link{tomstats}} with the event attribute.}
}
\description{
Specifies the statistic for an `event` effect in the tie-oriented model.
}
\details{
An `event` effect refers to an exogenous event attribute that affects the 
waiting time between events. The statistic at timepoint \emph{t} is for all 
dyads in the riskset equal to the attribute of the event at timepoint 
\emph{t}.
}
\examples{
history$work <- ifelse(history$setting == "work", 1, 0)
effects <- ~ event("work")
remstats(edgelist = history, tie_effects = effects, attributes = info)
tomstats(effects, edgelist = history)

}
\seealso{
\code{\link{FEtype}}
}
