% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{event}
\alias{event}
\title{event}
\usage{
event(x)
}
\arguments{
\item{x}{numeric vector with the attribute of the events. The length of this 
vector should be equal to the number of events in the edgelist}
}
\description{
Specifies the statistic for an event effect in the \code{formula} argument 
of \code{\link{remstats}}.
}
\details{
An event effect refers to an exogenous event attribute that affects the rate 
of interacting in events higher or lower on this attribute. The statistic at 
timepoint \emph{t} is for all dyads in the riskset equal to the attribute of 
the event at timepoint \emph{t}.
}
\examples{
data(history)
remstats(~ event(history$setting), edgelist = history)

}
