% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{event}
\alias{event}
\title{event}
\usage{
event(x, variableName = NULL)
}
\arguments{
\item{x}{vector with the event attribute}

\item{variableName}{optionally, a string indicating the variable name, used 
for the dimnames of the output statistics object}
}
\description{
Specifies the statistic for an `event` effect in the tie-oriented model.
}
\details{
An `event` effect refers to an exogenous event attribute that affects the 
waiting time between events. The statistic at timepoint \emph{t} is for all 
dyads in the riskset equal to the attribute of the event at timepoint 
\emph{t}.
}
\examples{
history$work <- ifelse(history$setting == "work", 1, 0)
effects <- ~ event("work")
remstats(edgelist = history, tie_effects = effects, attributes = info)
tomstats(effects, edgelist = history)

}
\seealso{
\code{\link{FEtype}}
}
