% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{inertia}
\alias{inertia}
\title{inertia}
\usage{
inertia(edgelist, riskset, weights, standardize)
}
\arguments{
\item{edgelist}{3-column edgelist (time, sender, receiver)}

\item{riskset}{2-column riskset (sender/actor 1, receiver/actor 2)}

\item{weights}{vector (length edgelist)}

\item{standardize}{logical}
}
\value{
matrix (time x dyad)
}
\description{
A function to compute the inertia effect.
}
\examples{
data(edgelistD)
out <- prepER(edgelist = edgelistD)
edgelist <- out$edgelist
riskset <- out$riskset
equal_weights <- rep(1, nrow(edgelist))
stat <- inertia(edgelist, riskset, equal_weights, standardize = FALSE)

}
