% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{inertia}
\alias{inertia}
\title{inertia}
\usage{
inertia(evls, riskset, weights, standardize)
}
\arguments{
\item{evls}{2-column edgelist (event, time) in relevent::rem format.}

\item{riskset}{2-column riskset (sender/actor 1, receiver/actor 2).}

\item{weights}{vector (length evls)}

\item{standardize}{logical}
}
\value{
matrix (time x dyad)
}
\description{
A function to compute the inertia effect.
}
\examples{
data(edgelistD)
out <- prepER(edgelist = edgelistD, directed = TRUE, type = FALSE, 
    riskset = NULL, actors = NULL)
el <- out$edgelist
rs <- out$riskset
evls <- prepEvls(el, rs, type = FALSE)
stat <- inertia(evls, rs, weights = rep(1, nrow(el)), FALSE)

}
