% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{inertia}
\alias{inertia}
\title{inertia}
\usage{
inertia(
  scaling = c("counts", "outdegreeSender", "standardize", "FrPSndSnd"),
  memory_value = Inf,
  with_type = FALSE,
  event_weights = NULL
)
}
\arguments{
\item{scaling}{the method for scaling the inertia statistic. Options are one 
of \code{"counts"} (default option, gives the raw counts of past (i,j) 
events at time t), \code{"outdegreeSender"} (in which raw counts are 
divided by the outdegree of the sender at time t, only available for 
directed events), and \code{"standardize"} (in which raw counts are 
standardized per time point).}

\item{memory_value}{numeric value. Specifies the time after which events are 
no longer included in the statistic count (default: all past events are 
considered). Note: make sure memory_value is specified in the same time unit 
as the time for the events in the edgelist.}

\item{with_type}{logical value. If TRUE, past (i,j) events of different 
types c are counted separately for every (i,j,c) event in the risk set. If 
FALSE (default), the type of the events is not considered in the counting of 
past (i,j) events. Note: with_type can be FALSE even when a distinction 
between different event types is made in the risk set but cannot be TRUE if 
a distinction between different event type is not made in the riskset.}

\item{event_weights}{vector with numeric values that indicate the intensity 
of the events in the edgelist. If event weights are supplied, the inertia 
statistic is equal to the sum of the intensity of past (i,j) events.}
}
\description{
Specifies the statistic for an inertia effect in the \code{formula} argument 
of \code{\link{remstats}}.
}
\details{
An inertia effect refers to the tendency for actors to repeat past 
interactions. The statistic at timepoint \emph{t} for dyad \emph{(i,j)} is 
equal to the number of \emph{(i,j)} events before timepoint \emph{t}. 
Optionally, a scaling method can be set with \code{scaling}, events that 
happened a certain number of time units ago can be disregarded in the count 
by setting \code{memory_value}, events of different types can be counted 
separately by setting \code{with_type = TRUE} or past events can be weighted 
by setting \code{event_weights}.
}
\examples{
data(history)
remstats(~ inertia(), edgelist = history)

}
