% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{inertiaMW}
\alias{inertiaMW}
\title{inertiaMW}
\usage{
inertiaMW(
  full_edgelist,
  window_edgelist,
  window_length,
  riskset,
  full_weights,
  standardize
)
}
\arguments{
\item{full_edgelist}{3-column edgelist (time, sender, receiver)}

\item{window_edgelist}{3-column edgelist (time, sender, receiver)}

\item{window_length}{numeric value.}

\item{riskset}{2-column riskset (sender/actor 1, receiver/actor 2).}

\item{full_weights}{vector (length full_edgelist).}

\item{standardize}{logical.}
}
\value{
matrix (time x dyad)
}
\description{
A function to compute the inertia effect.
}
\examples{
windows <- data.frame(start = seq(0, 900, 75), end = seq(100, 1000, 75))
data(edgelistU)
window_edgelist <- edgelistU[edgelistU$time > windows$start[2] & 
    edgelistU$time <= windows$end[2],]
out <- prepER(edgelist = edgelistU, directed = FALSE)
full_el <- out$edgelist
rs <- out$riskset
ac <- out$actors
out <- prepER(window_edgelist, directed = FALSE, actors = ac[,2])
window_el <- out$edgelist
stat <- inertiaMW(full_edgelist = full_el, window_edgelist = window_el, 
    window_length = 100, riskset = rs, 
    full_weights = rep(1, nrow(full_el)), standardize = FALSE)

}
