% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{maximum}
\alias{maximum}
\title{maximum}
\usage{
maximum(variable, attr_actors = NULL, scaling = c("none", "std"))
}
\arguments{
\item{variable}{string with the name of the column in the
\code{attr_actors} object for which the statistic has to be computed.}

\item{attr_actors}{optionally, an object of class
\code{\link[base]{data.frame}} that contains the attribute, see
'Details.'}

\item{scaling}{the method for scaling the statistic. Default is to not scale
the statistic. Alternatively, standardization of the
statistic per time point can be requested with "std".}
}
\description{
Specifies the statistic for a "maximum" effect in the tie-oriented model. A
"maximum" effect refers to an exogenous actor attribute that affects dyad
\emph{(i,j)}'s rate of interacting based on the maximum of the values of
actors \emph{i} and \emph{j} on this attribute.
}
\details{
The statistic at timepoint \emph{t} for dyad \emph{(i,j)} is equal to the
maximum of the values of actor \emph{i} and \emph{j} on the attribute at
timepoint \emph{t}.

The \code{attr_actors} object should be constructed as a
\code{\link[base]{data.frame}} where each row refers to the attribute value
of actor \emph{i} at timepoint \emph{t}:
\itemize{
 \item{\code{name}: the actors' name}
 \item{\code{time}: the time when the attribute values change (set to a
column with only zero's if the attribute does not vary over time).}
 \item{}{the third column contains the attribute that is called in the
specification of the "maximum" effect (i.e., the column name should
correspond to the string that is supplied to the \code{variable} argument)}
}
Note that it is possible to omit the attr_actors object in the call of
\code{maximum()} and, instead, supply it in the call of \code{remstats()} for
multiple exogenous effects.
}
\examples{
reh_tie <- remify::remify(history, model = "tie", directed = FALSE)
effects <- ~ maximum("extraversion")
remstats(reh = reh_tie, tie_effects = effects, attr_actors = info)

}
