% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepER.R
\name{prepER}
\alias{prepER}
\title{prepER}
\usage{
prepER(edgelist, riskset, directed, type)
}
\arguments{
\item{edgelist}{[matrix] or [dataframe], should minimally contain the time, 
sender/actor 1 and receiver/actor 2 in the first three columns respectively. 
If the riskset contains typed relational events, the fourth column should 
contain the event type.}

\item{riskset}{optional argument. If supplied, [matrix] or [dataframe], 
should minimally contain sender/actor 1 and receiver/actor 2 in the first 
two columns, respectively. If it contains typed relational events, the third 
column should contain the event type.}

\item{directed}{[logical], are relational events in the riskset directional 
(directed = TRUE) or undirectional (directed = FALSE).}

\item{type}{[logical], do relational events in the riskset consider an 
action type (type = TRUE) or not (type = FALSE).}
}
\value{
edgelist [matrix] with actor IDs that run from 1 to N and types that #' run from 1 to C.

riskset [matrix] with actor IDs that run from 1 to N and types that
run from 1 to C.
}
\description{
A helper function to prepare the edgelist and riskset for and within 
remstats().
}
\examples{
data(edgelistD)
out <- prepER(edgelistD, riskset = NULL, directed = TRUE, type = FALSE)
el <- out$edgelist
rs <- out$riskset

}
