% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{psABBY}
\alias{psABBY}
\title{psABBY}
\usage{
psABBY(consider_type = FALSE)
}
\arguments{
\item{consider_type}{logical, indicates whether to consider the event type
in determining which dyads create a pshift (TRUE) or not (FALSE, default).}
}
\description{
Specifies the statistic for a pshift AB-BY effect in the \code{effects}
argument of \code{\link{tomstats}}.
}
\details{
The AB-BY pshift effect refers to one of Gibson's (2003) dyadic
participation shifts. The AB-BY pshift refers to a tendency for turn
receiving (here, the next sender is the current receiver and the next
receiver is not in the current event). For each timepoint t, the psABBY
statistic is equal to one for the dyads that will create the participation
shift if they would occur in the edgelist at time t and equal to zero for
the dyads that will not create this participation shift. If consider_type is
set to TRUE, the type of the AB event and the type of the BY events have to
be equal. If it is set to FALSE, the participation shift is set to one for
every BY event, regardless of the event type. If multiple events in the
edgelist occur at the same time point, the order of these events determines
whether the p-shift is observed. Note that the AB-BY pshift is only defined
for directed events.
}
\examples{
reh_tie <- remify::remify(history, model = "tie")
effects <- ~ psABBY()
remstats(reh = reh_tie, tie_effects = effects)

}
\seealso{
\code{\link{psABBA}}, \code{\link{psABXA}}, \code{\link{psABXB}},
\code{\link{psABXY}} or \code{\link{psABAY}} for other dyadic participation
shifts.
}
