% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{psABXB}
\alias{psABXB}
\title{psABXB}
\usage{
psABXB(consider_type = FALSE, param = NULL)
}
\arguments{
\item{consider_type}{logical, indicates whether to consider the event type 
in determining which dyads create a pshift (TRUE) or not (FALSE, default).}

\item{param}{numeric value or function with time parameter to be used for 
generation of relational event histories with \code{\link{remulate}}.}
}
\description{
Specifies the statistic for a pshift AB-XB effect in the \code{effects} 
argument of \code{\link{tomstats}}.
}
\details{
The AB-XB pshift effect refers to one of Gibson's (2003) dyadic 
participation shifts. The AB-XB pshift refers to a tendency for turn 
usurping (here, the next sender is not in the current event and the next 
receiver is the current receiver). For each timepoint t, the psABXB 
statistic is equal to one for the dyads that will create the participation 
shift if they would occur in the edgelist at time t and equal to zero for 
the dyads that will not create this participation shift. If consider_type is 
set to TRUE, the type of the AB event and the type of the XB events have to 
be equal. If it is set to FALSE, the participation shift is set to one for 
every XB event, regardless of the event type. If multiple events in the 
edgelist occur at the same time point, the order of these events determines 
whether the p-shift is observed. Note that the AB-XB pshift is only defined 
for directed events.
}
\examples{
library(remify)
rehObject <- reh(history) 
effects <- ~ psABXB()
tomstats(effects, edgelist = rehObject)

}
\seealso{
\code{\link{psABBA}}, \code{\link{psABBY}}, \code{\link{psABXA}}, 
\code{\link{psABXY}} or \code{\link{psABAY}} for other dyadic participation 
shifts.
}
