% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{receive}
\alias{receive}
\title{receive}
\usage{
receive(variable, attributes = NULL, scaling = c("as.is", "std"))
}
\arguments{
\item{variable}{string with the name of the column in the 
\code{attributes} object for which the statistic has to be computed.}

\item{attributes}{optionally, an object of class 
\code{"\link[base]{data.frame}"} that contains the exogenous attributes, see 
'Details.'}

\item{scaling}{the method for scaling the statistic. Default is to not scale 
the statistic but keep it "as.is". Alternatively, standardization of the 
statistic per time point can be requested with "std".}
}
\description{
Specifies the statistic for a `receive` effect in the tie-oriented model or 
the receiver choice step of the actor-oriented model.
}
\details{
A receive effect refers to an exogenous actor attribute that affects actor 
i's rate of receiving events. The statistic at timepoint \emph{t} is equal 
to the value of the exogenous attribute for actor \emph{i} at time \emph{t} 
for all dyads in the riskset that have actor \emph{i} as receiver. Note that 
a receive effect is only defined for directed relational events.

The \code{attributes} object should be constructed as follows: Each row 
refers to the attribute value of actor \emph{i} at timepoint \emph{t}. An 
`id` column is required that contains the actor id (corresponding to the 
actor id's in the edgelist). A `time` column is required that contains the 
time when attributes change (set to zero if none of the attributes vary over 
time). Subsequent columns contain the attributes that are called in the 
specifications of exogenous statistics. Alternatively, a dataframe with 
attributes for all exogenous effects can be defined in the \code{attributes} 
argument of \code{\link{remstats}}, \code{\link{tomstats}} or 
\code{\link{aomstats}}.
}
\examples{
library(remify)
rehObject <- reh(history)
effects <- ~ receive("extraversion") 
remstats(edgelist = rehObject, tie_effects = effects, attributes = info)
remstats(edgelist = rehObject, receiver_effects = effects, attributes = info)
tomstats(effects, edgelist = rehObject, attributes = info)
aomstats(receiver_effects = effects, edgelist = rehObject, attributes = info)

}
