% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{receive}
\alias{receive}
\title{receive}
\usage{
receive(variable, attributes = NULL, scaling = c("as.is", "std"))
}
\arguments{
\item{variable}{string with the name of the column in the
\code{attributes} object for which the statistic has to be computed.}

\item{attributes}{optionally, an object of class
\code{\link[base]{data.frame}} that contains the attribute, see
'Details.'}

\item{scaling}{the method for scaling the statistic. Default is to not scale
the statistic but keep it "as.is". Alternatively, standardization of the
statistic per time point can be requested with "std".}
}
\description{
Specifies the statistic for a "receive" effect in the tie-oriented model or
the receiver choice step of the actor-oriented model. A "receive" effect
refers to an exogenous actor attribute that affects actor \emph{i}'s rate of
receiving events.
}
\details{
The statistic at timepoint \emph{t} is equal to the value of the exogenous
attribute for actor \emph{i} at time \emph{t} for all dyads in the riskset
that have actor \emph{i} as receiver. Note that a "receive" effect is only
defined for directed relational events.

The \code{attributes} object should be constructed as a
\code{\link[base]{data.frame}} where each row refers to the attribute value
of actor \emph{i} at timepoint \emph{t}:
\itemize{
 \item{\code{id}: the actors' id}
 \item{\code{time}: the time when the attribute values change (set to a
column with only zero's if the attribute does not vary over time).}
 \item{}{the third column contains the attribute that is called in the
specification of the "receive" effect (i.e., the column name should
correspond to the string that is supplied to the \code{variable} argument)}
}
Note that it is possible to omit the attributes object in the call of
\code{receive()} and, instead, supply it in the call of \code{remstats()}
for multiple exogenous effects.
}
\examples{
reh_tie <- remify::remify(history, model = "tie")
effects <- ~ receive("extraversion")
remstats(reh = reh_tie, tie_effects = effects, attributes = info)

reh_actor <- remify(history, model = "actor")
remstats(reh = reh_actor, receiver_effects = effects, attributes = info)

}
