% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{recencyReceiveSender}
\alias{recencyReceiveSender}
\title{recencyReceiveSender}
\usage{
recencyReceiveSender(consider_type = FALSE)
}
\arguments{
\item{consider_type}{logical, indicates whether to compute the recency
separately for each event type (TRUE) or regardless of event types (FALSE, 
default).}
}
\description{
Specifies the statistic for a recency receive of sender effect in the 
\code{effects} argument of \code{\link{tomstats}} or the 
\code{rateEffects} argument of \code{\link{aomstats}}.
}
\details{
The recencyReceiveSender effect refers to a recency statistic similar to 
what is described in Vu et al. (2017) and Mulder and Leenders (2019). For 
each timepoint t, for directed dyad (i,j) the statistic is equal to 1/(the 
time that has past since sender i was last active as receiver + 1). Note 
that the 'recencyReceiveSender' effect is only defined for directed events.
}
\examples{
effects <- ~ recencyReceiveSender()
tomstats(effects, edgelist = history)
aomstats(rateEffects = effects, edgelist = history)

}
\seealso{
\code{\link{rrankSend}}, \code{\link{rrankReceive}}, 
\code{\link{recencySendSender}}, \code{\link{recencySendReceiver}}, 
\code{\link{recencyReceiveReceiver}} and \code{\link{recencyContinue}} for 
other type of recency effects
}
