% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{reciprocity}
\alias{reciprocity}
\title{reciprocity}
\usage{
reciprocity(
  scaling = c("counts", "indegreeSender", "standardize"),
  memory_value = Inf,
  with_type = FALSE,
  event_weights = NULL
)
}
\arguments{
\item{scaling}{the method for scaling the reciprocity statistic. Options are 
one of \code{"counts"} (default option, gives the raw counts of past (j,i) 
events at time t), \code{"indegreeSender"} (in which raw counts are 
divided by the indegree of the sender at time t), and \code{"standardize"} 
(in which raw counts are standardized per time point).}

\item{memory_value}{numeric value. Specifies the time after which events are 
no longer included in the statistic count (default: all past events are 
considered). Note: make sure memory_value is specified in the same time unit 
as the time for the events in the edgelist.}

\item{with_type}{logical value. If TRUE, past (j,i) events of different 
types c are counted separately for every (i,j,c) event in the risk set. If 
FALSE (default), the type of the events is not considered in the counting of 
past (j,i) events. Note: with_type can be FALSE even when a distinction 
between different event types is made in the risk set.}

\item{event_weights}{vector with numeric values that indicate the intensity 
of the events in the edgelist. If event weights are supplied, the 
reciprocity statistic is equal to the sum of the intensity of past (j,i)
events.}
}
\description{
Specifies the statistic for a reciprocity effect in the \code{formula} 
argument of \code{\link{remstats}}.
}
\details{
A reciprocity effect refers to the tendency for actors to reciprocate past 
interactions. The statistic at timepoint \emph{t} for dyad \emph{(i,j)} is 
equal to the number of \emph{(j,i)} events before timepoint \emph{t}. 
Optionally, a scaling method can be set with \code{scaling}, events that 
happened a certain number of time units ago can be disregarded in the count 
by setting \code{memory_value}, events of different types can be counted 
separately by setting \code{with_type = TRUE} or past events can be weighted 
by setting \code{event_weights}.

Note that by scaling the reciprocity count by the indegree of the sender, 
the statistic refers to the fraction of messages received by actor i that 
were received from actor j. If actor i hasn't received any messages yet it 
can be assumed that actor i is equally likely to receive a message from 
every actor and the statistic is set equal to 1/(n-1), where n refers to the 
number of actors. The resulting statistic is similar to the "FrRecSnd" 
statistic in the R package relevent.
}
\examples{
data(history)
remstats(~ reciprocity(), edgelist = history)

}
