% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remstats.R
\name{remStats}
\alias{remStats}
\title{remStats}
\usage{
remStats(edgelist, effects, directed = TRUE, type = FALSE,
  timing = "interval", riskset = NULL, actors = NULL,
  weights = NULL)
}
\arguments{
\item{edgelist}{[matrix] or [dataframe], should minimally contain the time, 
sender/actor 1 and receiver/actor 2 in the first three columns respectively. 
If the riskset contains typed relational events, the fourth column should 
contain the event type.}

\item{effects}{[character vector], indicates the effects that are requested.}

\item{directed}{[logical], are relational events in the riskset directional 
(directed = TRUE, default) or undirectional (directed = FALSE).}

\item{type}{[logical], do relational events in the riskset consider an 
action type (type = TRUE) or not (type = FALSE, default).}

\item{timing}{[character value], indicates whether the full likelihood 
(timing = "interval", default) or ordinal likelihod (timing = "ordinal") 
will be used for estimation. If interval timing, a baseline statistic is 
added to the statistic array.}

\item{riskset}{optional; [matrix] or [dataframe], should minimally contain 
sender/actor 1 and receiver/actor 2 in the first two columns, respectively. 
If it contains typed relational events, the third column should contain the 
event type. If a riskset is not supplied, it is assumed that all possible 
actors (and action types) are observed in the edgelist.}

\item{actors}{optional; [vector], if supplied, should contain all actors 
that can potentially interact. Used to create the riskset.}

\item{weights}{optional; [vector], if supplied, should be of length edgelist #' and contain the weights for the events in the edgelist (to compute 
inertia_weighted)}
}
\value{
statistics [array], with three dimensions: timepoint x riskset x 
statistic.

edgelist [matrix] with actor IDs that run from 1 to N and types that #' run from 1 to C.

riskset [matrix], with actor IDs that run from 1 to N and types that
run from 1 to C.

evls [matrix], edgelist transformed in the format that is required 
for estimation by relevent::rem() with in the first column the relational 
event ID and in the second column the time.

actors [vector], all unique actor IDs
}
\description{
A function to compute statistics for a relational event sequence.
}
\examples{
data(edgelistD)
effects <- c("inertia", "indegree_receiver", "outdegree_sender")
out <- remStats(edgelistD, effects)
fit <- relevent::rem(out$evls, out$statistics)
summary(fit)

}
