% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remstats.R
\name{remStats}
\alias{remStats}
\title{remStats}
\usage{
remStats(edgelist, effects, riskset = NULL, directed = TRUE,
  type = FALSE)
}
\arguments{
\item{edgelist}{[matrix] or [dataframe], should minimally contain the time, 
sender/actor 1 and receiver/actor 2 in the first three columns respectively. 
If the riskset contains typed relational events, the fourth column should 
contain the event type.}

\item{effects}{[character vector], indicates the effects that are requested.}

\item{riskset}{optional; [matrix] or [dataframe], should minimally contain 
sender/actor 1 and receiver/actor 2 in the first two columns, respectively. 
If it contains typed relational events, the third column should contain the 
event type. If a riskset is not supplied, it is assumed that all possible 
actors (and action types) are observed in the edgelist.}

\item{directed}{[logical], are relational events in the riskset directional 
(directed = TRUE, default) or undirectional (directed = FALSE).}

\item{type}{[logical], do relational events in the riskset consider an 
action type (type = TRUE) or not (type = FALSE, default).}
}
\value{
statistics [array], with three dimensions: timepoint x riskset x 
statistic.

evls [matrix], edgelist transformed in the format that is required 
for estimation by relevent::rem() with in the first column the relational 
event ID and in the second column the time.
}
\description{
A function to compute statistics for a relational event sequence.
}
