% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remstats.R
\name{remstats}
\alias{remstats}
\title{remstats}
\usage{
remstats(
  formula,
  edgelist,
  directed = TRUE,
  with_type = FALSE,
  riskset = NULL,
  actors = NULL,
  types = NULL,
  start = NULL,
  stop = NULL
)
}
\arguments{
\item{formula}{an object of class \code{"\link[stats]{formula}"} (or one 
that can be coerced to that class): a symbolic description of statistics to 
be computed. The details of the specification of the statistics and an 
overview of the available statistics are given under 'Details'.}

\item{edgelist}{an object of class \code{"\link[base]{matrix}"} or 
\code{"\link[base]{data.frame}"} that contains the relational event history. 
Each row in the edgelist should refer to one event. The first column should 
refer to the timepoint or order of the event, the second column to the 
sender (or first actor) and the third column to the receiver (or second 
actor). The fourth column may refer to the type of the event, this column is 
only used when \code{with_type = TRUE}.}

\item{directed}{Logical value. Indicates whether events in the edgelist are 
directed (\code{directed = TRUE}, default) or undirected 
(\code{directed = FALSE}).}

\item{with_type}{Logical value. Indicates whether event types are considered 
in the dependent variable (\code{with_type = TRUE}) or not 
(\code{with_type = FALSE}, default).}

\item{riskset}{an object of class \code{"\link[base]{matrix}"} or 
\code{"\link[base]{data.frame}"} that contains the riskset. The first column 
should refer to the sender/actor1, the second column to the receiver/actor2 
and a third column may refer to the event type. Can be supplied to indicate 
a non-standard riskset.}

\item{actors}{Vector with actor id's. Should be supplied when not all actors 
that can interact are observed in the edgelist.}

\item{types}{Vector with event type id's. Should be supplied when not all 
types that can occur are observed in the edgelist.}

\item{start}{Integer value. Indicates the first row in the edgelist for 
which statistics need to be computed. Can be used to compute statistics for 
a subpart of the relational event history but based on the whole relational 
event history. If nothing is indicated, statistics will be computed starting 
from the first row of the edgelist.}

\item{stop}{Integer value. Indicates the last row in the edgelist for 
which statistics need to be computed. If nothing is indicated, statistics 
will be computed stopping at the last row of the edgelist.}
}
\description{
Computes statistics for relational event history data.
}
\details{
The statistics to be computed are defined symbolically and should be in the 
form \code{~ statistics}. The statistics are separated by + operators. 
Interactions between two statistics can be included with * or : operators. 

A list of available statistics follows: 
\itemize{
 \item \code{\link{baseline}()}
 \item \code{\link{send}()}
 \item \code{\link{receive}()}
 \item \code{\link{same}()}
 \item \code{\link{difference}()}
 \item \code{\link{average}()}
 \item \code{\link{minimum}()}
 \item \code{\link{maximum}()}
 \item \code{\link{equate}()}
 \item \code{\link{event}()}
 \item \code{\link{inertia}()}
 \item \code{\link{reciprocity}()}
 \item \code{\link{indegreeSender}()}
 \item \code{\link{indegreeReceiver}()}
 \item \code{\link{outdegreeSender}()}
 \item \code{\link{outdegreeReceiver}()}
 \item \code{\link{totaldegreeSender}()}
 \item \code{\link{totaldegreeReceiver}()}
 \item \code{\link{otp}()}
 \item \code{\link{itp}()}
 \item \code{\link{osp}()}
 \item \code{\link{isp}()}
 \item \code{\link{sp}()}
 \item \code{\link{spUnique}()}
 \item \code{\link{psABBA}()}
 \item \code{\link{psABBY}()}
 \item \code{\link{psABXA}()}
 \item \code{\link{psABXB}()}
 \item \code{\link{psABXY}()}
 \item \code{\link{psABAY}()}
 \item \code{\link{rrankSend}()}
 \item \code{\link{rrankReceive}()}
 \item \code{\link{recenySender}()}
 \item \code{\link{recenyReceiver}()}
 \item \code{\link{recenyContinue}()}
}
}
\examples{
data(history)
data(info)
form <- ~ baseline() + inertia():send("extraversion", info)
remstats(form, edgelist = history)

}
