% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remstats.R
\name{remstats}
\alias{remstats}
\title{remstats}
\usage{
remstats(
  formula,
  edgelist,
  actors = NULL,
  types = NULL,
  directed = TRUE,
  ordinal = FALSE,
  origin = NULL,
  omit_dyad = NULL,
  start = NULL,
  stop = NULL
)
}
\arguments{
\item{formula}{an object of class \code{"\link[stats]{formula}"} (or one 
that can be coerced to that class): a symbolic description of the requested 
statistics, see 'Details'.}

\item{edgelist}{an object of class \code{"\link[base]{data.frame}"} or 
\code{"\link[base]{matrix}"} characterizing the relational event history 
sorted by time with columns `time`, `actor1`, `actor2` and optionally `type` 
and `weight`. Alternatively, an object of class \code{"\link[remify]{reh}"}.}

\item{actors}{vector of actors that may be observed interacting in the network. If \code{NULL}, actor names will be drawn from the input edgelist.}

\item{types}{vector of event types that may occur in the network. If \code{NULL}, type names will be drawn from the input edgelist.}

\item{directed}{logical value indicating whether dyadic events are directed (\code{TRUE}) or undirected (\code{FALSE}).}

\item{ordinal}{logical value indicating whether only the order of events matters in the model (\code{TRUE}) or also the waiting time must be considered in the model (\code{FALSE}).}

\item{origin}{time point since which when events could occur (default is \code{NULL}). If it is defined, it must have the same class of the time column in the input edgelist.}

\item{omit_dyad}{list of lists of two elements: `time`, that is a vector of the time points which to omit dyads from, `dyad`, which is a \code{"\link[base]{data.frame}"} where dyads to be omitted are supplied.}

\item{start}{integer value that indicates the first row in the edgelist for 
which statistics need to be computed. Can be used to compute statistics for 
a subpart of the relational event history based on the whole past relational 
event history. If nothing is indicated, statistics will be computed starting 
from the first row of the edgelist.}

\item{stop}{integer value that indicates the last row in the edgelist for 
which statistics need to be computed. If nothing is indicated, statistics 
will be computed stopping at the last row of the edgelist.}
}
\description{
Computes statistics for relational event history data.
}
\details{
The statistics to be computed are defined symbolically and should be in the 
form \code{~ statistics}. The statistics are separated by + operators. 
Interactions between two statistics can be included with * or : operators. 

A list of available statistics follows: 
\itemize{
 \item \code{\link{baseline}()}
 \item \code{\link{send}()}
 \item \code{\link{receive}()}
 \item \code{\link{tie}()}
 \item \code{\link{same}()}
 \item \code{\link{difference}()}
 \item \code{\link{average}()}
 \item \code{\link{minimum}()}
 \item \code{\link{maximum}()}
 \item \code{\link{equate}()}
 \item \code{\link{event}()}
 \item \code{\link{inertia}()}
 \item \code{\link{reciprocity}()}
 \item \code{\link{indegreeSender}()}
 \item \code{\link{indegreeReceiver}()}
 \item \code{\link{outdegreeSender}()}
 \item \code{\link{outdegreeReceiver}()}
 \item \code{\link{totaldegreeSender}()}
 \item \code{\link{totaldegreeReceiver}()}
 \item \code{\link{otp}()}
 \item \code{\link{itp}()}
 \item \code{\link{osp}()}
 \item \code{\link{isp}()}
 \item \code{\link{sp}()}
 \item \code{\link{spUnique}()}
 \item \code{\link{psABBA}()}
 \item \code{\link{psABBY}()}
 \item \code{\link{psABXA}()}
 \item \code{\link{psABXB}()}
 \item \code{\link{psABXY}()}
 \item \code{\link{psABAY}()}
 \item \code{\link{rrankSend}()}
 \item \code{\link{rrankReceive}()}
 \item \code{\link{recencySender}()}
 \item \code{\link{recencyReceiver}()}
 \item \code{\link{recencyContinue}()}
}
}
\examples{
data(history)
data(info)
remstats(~ baseline() + inertia():send("extraversion", info), 
 edgelist = history)

}
