% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{remStatsC}
\alias{remStatsC}
\title{remStatsC}
\usage{
remStatsC(effects, edgelist, riskset, evls, actors)
}
\arguments{
\item{effects}{integer vector (effects)}

\item{edgelist}{3-column edgelist (time, sender, receiver)}

\item{riskset}{2-column riskset (sender/actor 1, receiver/actor 2)}

\item{evls}{2-column edgelist (event, time) in relevent::rem format}

\item{actors}{vector with numeric actor IDs (correspod to edgelist, riskset)}
}
\value{
statistics 3-dimensional array (event time x risk set entry x 
    statistic)
}
\description{
A function to compute statistics and combine the statistics in an array
prepared for estimation of a REM with relevent::rem()
}
