% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{rrankReceive}
\alias{rrankReceive}
\title{rrankReceive}
\usage{
rrankReceive(with_type = FALSE)
}
\arguments{
\item{with_type}{logical value. If TRUE, ranks are created taking event 
types into account. If FALSE (default), ranks are created regardless of the 
event types of the dyads in the riskset.}
}
\description{
Specifies the statistic for a recency rank receive effect in the 
\code{formula} argument of \code{\link{remstats}}.
}
\details{
The rrankSend effect refers to a rank-based recency effect, as in section 2.
2.5 of Butts (2008). For each timepoint t, for directed dyad (i,j) the 
statistic is equal to the inverse of the rank of receiver j among the actors 
from which sender i has most recently received past events
}
\examples{
data(history)
remstats(~ rrankReceive(), edgelist = history)

}
\seealso{
\code{\link{rrankSend}}
}
