% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{rrankSend}
\alias{rrankSend}
\alias{recency}
\alias{recencyRank}
\alias{rrank}
\title{rrankSend}
\usage{
rrankSend(with_type = FALSE)
}
\arguments{
\item{with_type}{logical value. If TRUE, ranks are created taking event 
types into account. If FALSE (default), ranks are created regardless of the 
event types of the dyads in the riskset.}
}
\description{
Specifies the statistic for a recency rank send effect in the 
\code{formula} argument of \code{\link{remstats}}.
}
\details{
The rrankSend effect refers to a rank-based recency effect, as in section 2.
2.5 of Butts (2008). For each timepoint t, for directed dyad (i,j) the 
statistic is equal to the inverse of the rank of receiver j among the actors 
to which sender i has most recently send past events.
}
\examples{
data(history)
remstats(~ rrankSend(), edgelist = history)

}
\seealso{
\code{\link{rrankReceive}}
}
