% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{rrankSend}
\alias{rrankSend}
\alias{recencyRank}
\alias{rrank}
\title{rrankSend}
\usage{
rrankSend(consider_type = FALSE)
}
\arguments{
\item{consider_type}{logical, indicates whether to determine the rank 
separately for each event type (TRUE) or regardless of event types (FALSE, 
default).}
}
\description{
Specifies the statistic for a recency rank send effect in the 
\code{effects} argument of \code{\link{tomstats}} or the 
\code{choiceEffects} argument of \code{\link{aomstats}}.
}
\details{
The rrankSend effect refers to a rank-based recency effect, as described in 
section 2.2.5 of Butts (2008). For each timepoint t, for directed dyad (i,j) 
the statistic is equal to the inverse of the rank of receiver j among the 
actors to which sender i has most recently send past events. Note that the 
'rrankSend' effect is only defined for directed events.
}
\examples{
effects <- ~ rrankSend()
tomstats(effects, edgelist = history)
aomstats(choiceEffects = effects, edgelist = history)

}
\seealso{
\code{\link{rrankReceive}}, \code{\link{recencySendSender}}, 
\code{\link{recencySendReceiver}}, \code{\link{recencyReceiveSender}}, 
\code{\link{recencyReceiveReceiver}} and \code{\link{recencyContinue}} for 
other type of recency effects
}
