% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{same}
\alias{same}
\title{same}
\usage{
same(variables, covariates)
}
\arguments{
\item{variables}{character vector: names of one or more columns in the 
\code{covariates} object for which the statistic has to be computed.}

\item{covariates}{an object of class \code{"\link[base]{data.frame}"} that 
contains the exogenous covariates. Each row should refer to one actor. The 
first column should contain the actor id (corresponding to the actor id's in 
the edgelist). The second column refers to the time when covariates change 
(set to zero if none of the covariates vary over time). Subsequent columns 
contain the covariates that are called in the specifications of exogenous 
statistics.}
}
\description{
Specifies the statistic for a same effect in the \code{formula} argument 
of \code{\link{remstats}}.
}
\details{
A same effect refers to an exogenous actor attribute that affects dyad 
\emph{(i,j)}'s rate of interacting based on whether they have the same value 
(or not) on this attribute. The statistic at timepoint \emph{t} is equal to 
one for dyads \emph{(i,j)} that have the same value on the attribute at 
timepoint \emph{t} and equal to 0 for dyads that do not have the same value.
}
\examples{
data(history)
data(info)
remstats(~ same(c("age", "sex"), info), edgelist = history)

}
