% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{same}
\alias{same}
\title{same}
\usage{
same(variable, attributes = NULL)
}
\arguments{
\item{variable}{string with the name of the column in the 
\code{attributes} object for which the statistic has to be computed.}

\item{attributes}{optionally, an object of class 
\code{"\link[base]{data.frame}"} that contains the exogenous attributes (see 
details).}
}
\description{
Specifies the statistic for a `same` effect in the tie-oriented model or the 
receiver choice step of the actor-oriented model.
}
\details{
A `same` effect refers to an exogenous actor attribute that affects dyad 
\emph{(i,j)}'s rate of interacting (tie-oriented model) or actor \emph{j}'s  
probability of being chosen as a receiver for the event send by the active 
sender \emph{i} at time \emph{t} (actor-oriented model) based on whether 
actors \emph{i} and \emph{j} have the same value (or not) on this attribute. 
The statistic at timepoint \emph{t} is equal to one for dyads \emph{(i,j)} 
that have the same value on the attribute at timepoint \emph{t} 
(tie-oriented model) or one for receivers \emph{j} that have the same value 
on the attribute as the active sender \emph{i} at timepoint \emph{t} 
(actor-oriented model) and equal to 0 for dyads resp. receivers that do not 
have the same value. 

The \code{attributes} object should be constructed as follows: Each row 
refers to the attribute value of actor \emph{i} at timepoint \emph{t}. An 
`id` column is required that contains the actor id (corresponding to the 
actor id's in the edgelist). A `time` column is required that contains the 
time when attributes change (set to zero if none of the attributes vary over 
time). Subsequent columns contain the attributes that are called in the 
specifications of exogenous statistics. Alternatively, a dataframe with 
attributes for all exogenous effects can be defined in the \code{attributes} 
argument of \code{\link{remstats}}, \code{\link{tomstats}} or 
\code{\link{aomstats}}.
}
\examples{
library(remify)
rehObject <- reh(history)
effects <- ~ same("age") 
remstats(edgelist = rehObject, tie_effects = effects, attributes = info)
remstats(edgelist = rehObject, receiver_effects = effects, attributes = info)
tomstats(effects, edgelist = rehObject, attributes = info)
aomstats(receiver_effects = effects, edgelist = rehObject, attributes = info)

}
