% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{sp}
\alias{sp}
\title{sp}
\usage{
sp(scaling = c("as.is", "std"), consider_type = FALSE)
}
\arguments{
\item{scaling}{the method for scaling the triad statistic. Default is to not
scale the statistic but keep the raw 'counts'. Alternatively,
standardization of the raw counts per time point can be requested
with 'std'.}

\item{consider_type}{logical, indicates whether to count the shared partners
separately for each event type (TRUE) or sum across different event
types (FALSE, default).}
}
\description{
Specifies the statistic for a shared partners effect in the \code{effects}
argument of \code{\link{tomstats}}.
}
\details{
A shared partners effect refers to the tendency of dyads to interact if they
have more past shared partners between them. The statistic at timepoint
\emph{t} for dyad \emph{(i,j)} is equal to the minimum of past undirected
\emph{(i,h)}, \emph{(j,h)} events, summed over all actors \emph{h}.
Optionally, a scaling method can be set with \code{scaling}. Note that the
`shared partners' effect is only defined for undirected events.
}
\examples{
library(remify)
rehObject <- reh(history, directed = FALSE)
effects <- ~ sp()
tomstats(effects, reh = rehObject)

}
\seealso{
\code{\link{spUnique}} for another type of triadic effect for
undirected relational events and \code{\link{otp}}, \code{\link{itp}},
\code{\link{osp}}, or \code{\link{isp}} for triadic effects for directed
relational events.
}
