% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_stats.R
\name{stack_stats}
\alias{stack_stats}
\title{Stack remstats for model fitting}
\usage{
stack_stats(stats, reh)
}
\arguments{
\item{stats}{A \code{remstats} or \code{tomstats} object containing
statistics for the (active) risk set.}

\item{reh}{A \code{remify} object containing event ordering, risk sets,
and inter-event times.}
}
\value{
A list with elements:
\describe{
  \item{remstats_stack}{A data frame in stacked (long) format containing the
    response \code{obs}, the statistics, and the offset \code{log_intevent}.}
  \item{subset}{Integer vector of length two giving the first and last
    event indices used for computing \code{stats}.}
  \item{D}{}{Number of included dyads.}
  \item{E}{}{Number of included events}
}
}
\description{
Stack a \code{remstats} object (\code{tomstats}; \code{aomstats}) into long format for
model fitting. Each row corresponds to one dyad in the risk set at one
event time, with a response indicating the number of observed events for
that dyad at that time and an offset given by the log inter-event time.
}
