% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{tie}
\alias{tie}
\title{tie}
\usage{
tie(x, variableName = NULL, scaling = c("as.is", "std"))
}
\arguments{
\item{x}{a matrix with attribute information, rows and columns should refer 
to actors in the edgelist}

\item{variableName}{optionally, a string indicating the variable name, used 
for the dimnames of the output statistics object}

\item{scaling}{the method for scaling the statistic. Default is to not scale 
the statistic but keep it "as.is". Alternatively, standardization of the 
statistic per time point can be requested with "std".}
}
\description{
Specifies the statistic for a `tie` effect in the tie-oriented model or the 
receiver choice step of the actor-oriented model.
}
\details{
A tie effect refers to an exogenous dyad attribute that affects dyad 
\emph{(i,j)}'s rate of interacting. The statistic is equal to the value of 
the exogenous attribute for dyad \emph{(i,j)}.
}
\examples{
library(remify)
data(info)
actors <- unique(info$id)
age <- info[match(actors, info$id), "age"]
bothOld <- sapply(1:length(actors), function(i) {
 sapply(1:length(actors), function(j) {
     ifelse(age[i] == 1 & age[j] == 1 & i != j, 1, 0)
 })})
rehObject <- reh(history)
effects <- ~ tie(bothOld, variableName = "both.old")
remstats(edgelist = rehObject, tie_effects = effects, attributes = info)
remstats(edgelist = rehObject, receiver_effects = effects, attributes = info)
tomstats(effects, edgelist = rehObject)
aomstats(receiver_effects = effects, edgelist = rehObject)

}
