% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tomstats.R
\name{tomstats}
\alias{tomstats}
\title{tomstats}
\usage{
tomstats(
  effects,
  reh,
  attributes = NULL,
  memory = c("full", "window", "decay", "interval"),
  memory_value = NA,
  start = 1,
  stop = Inf,
  adjmat = NULL,
  get_adjmat = FALSE
)
}
\arguments{
\item{effects}{an object of class \code{"\link[stats]{formula}"} (or one
that can be coerced to that class): a symbolic description of the effects in
the model for which statistics are computed, see 'Details' for the available
effects and their corresponding statistics}

\item{reh}{an object of class \code{"\link[remify]{remify}"} characterizing 
the relational event history.}

\item{attributes}{optionally, an object of class
\code{"\link[base]{data.frame}"} that contains the exogenous attributes (see
Details).}

\item{memory}{The memory to be used. See `Details'.}

\item{memory_value}{Numeric value indicating the memory parameter. See
`Details'.}

\item{start}{an optional integer value, specifying the index of the first
event in the relational event history for which statistics must be computed
(see 'Details')}

\item{stop}{an optional integer value, specifying the index of the last
event in the relational event history for which statistics muts be computed
(see 'Details')}

\item{adjmat}{optionally, a previously computed adjacency matrix with on the
rows the time points and on the columns the risk set entries}

\item{get_adjmat}{whether the adjmat computed by tomstats should be
outputted as an attribute of the statistics.}
}
\value{
\code{statistics } array with the computed statistics, where rows
refer to time points, columns refer to potential relational event (i.e.,
potential edges) in the risk set and slices refer to statistics
}
\description{
Computes statistics for modeling relational event history data
with Butts' (2008) relational event model.
}
\section{Effects}{

The statistics to be computed are defined symbolically and should be
supplied to the \code{effects} argument in the form \code{~ effects}. The
terms are separated by + operators. For example:
\code{effects = ~ inertia() + otp()}. Interactions between two effects
can be included with * operators. For example:
\code{effects = ~ inertia()*otp()}. A list of available effects can be
obtained with \code{\link{tie_effects}()}.

The majority of the statistics can be scaled in some way, see
the documentation of the \code{scaling} argument in the separate effect
functions for more information on this.

The majority of the statistics can account for the event type
included as a dependent variable, see the documentation of the
\code{consider_type} argument in the separate effect functions for more
information on this.

Note that events in the relational event history can be directed or
undirected. Some statistics are only defined for either directed or
undirected events (see the documentation of the statistics). Note that
undirected events are only available for the tie-oriented model.
}

\section{Attributes}{

For the computation of the \emph{exogenous} statistics an attributes object
with the exogenous covariate information has to be supplied to the
\code{attributes} argument in either \code{remstats()} or in the separate
effect functions supplied to the \code{..._effects} arguments (e.g., see
\code{\link{send}}). This \code{attributes} object should be constructed as
follows: A dataframe with rows refering to the attribute value of actor
\emph{i} at timepoint \emph{t}. A `name` column is required that contains the
actor name (corresponding to the actor names in the relational event
history). A `time` column is required that contains the time when attributes
change (set to zero if none of the attributes vary over time). Subsequent
columns contain the attributes that are called in the specifications of
exogenous statistics (column name corresponding to the string supplied to
the \code{variable} argument in the effect function). Note that the
procedure for the exogenous effects `tie' and `event' deviates from this,
here the exogenous covariate information has to be specified in a different
way, see \code{\link{tie}} and \code{\link{event}}.
}

\section{Memory}{

The default `memory` setting is `"full"`, which implies that at each time
point $t$ the entire event history before $t$ is included in the computation
of the statistics. Alternatively, when `memory` is set to `"window"`, only
the past event history within a given time window is considered (see
Mulders & Leenders, 2019). This length of this time window is set by the
`memory_value` parameter. For example, when `memory_value = 100` and `memory
= "window"`, at time point $t$ only the past events that happened at most
100 time units ago are included in the computation of the statistics.
A third option is to set `memory` to `"interval"`. In this case, the past
event history within a given time interval is considered. For example, when
`"memory_value" = c(50, 100)` and `memory = "window"`, at time point $t$
only the past events tha happened between 50 and 100 time units ago are
included in the computation of the statistics. Finally, the fourth option is
to set `memory` to `"decay"`. In this case, the weight of the past event in
the computation of the statistics depend on the elapsed time between $t$ and
the past event. This weight is determined based on an exponential decay
function with half-life parameter `memory_value` (see Brandes et al., 2009).
}

\section{Event weights}{

Note that if the relational event history contains a column that is named
``weight'', it is assumed that these affect the endogenous statistics. These
affect the computation of all endogenous statistics with a few exceptions
that follow logically from their definition (e.g., the recenyContinue
statistic does depend on time since the event and not on event weights).
}

\section{Subset of the relational event history}{

Optionally, statistics can be computed for a slice of the relational event
sequence - but based on the entire history. This is achieved by setting the
start and stop values equal to the index of the first and last event for
which statistics are requested. For example, start = 5 and stop = 5 computes
the statistics for only the 5th event in the relational event sequence,
based on the history that consists of events 1-4.
}

\section{Adjacency matrix}{

Optionally, a previously computed adjacency matrix can be supplied. Note
that the endogenous statistics will be computed based on this adjacency
matrix. Hence, supplying a previously computed adjacency matrix can reduce
computation time but the user should be absolutely sure the adjacency matrix
is accurate.
}

\examples{
library(remstats)
reh <- remify::remify(edgelist = history, model = "tie")
effects <- ~ inertia():send("extraversion") + otp()
tomstats(effects, reh = reh, attributes = info)

}
\references{
Butts, C. T. (2008). A relational event framework for social
action. Sociological Methodology, 38(1), 155–200.
\url{https://doi.org/10.1111/j.1467-9531.2008.00203.x}
}
