% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{totaldegreeDyad}
\alias{totaldegreeDyad}
\title{totaldegreeDyad}
\usage{
totaldegreeDyad(scaling = c("as.is", "prop", "std"))
}
\arguments{
\item{scaling}{the method for scaling the degree statistic. Default is
to not scale the statistic but keep the raw counts 'as.is'. Alternatively,
scaling of the raw degree counts by two times the number of past events at
time t can be requested with 'prop' or standardization of the raw degree
counts per time point can be requested with 'std'.}
}
\description{
Specifies the statistic for a `totaldegreeDyad` effect in the tie-oriented
model.
}
\details{
A totaldegreeDyad effect refers to the tendency for dyads to increase their
interaction rate if the total degree of the two actors in the pair
increases. The statistic at timepoint \emph{t} for dyad \emph{(i,j)} is
equal to the sum of the event before timepoint \emph{t} that involved at
least one actor of the pair \emph{(i,j)}.

Optionally, a scaling method can be set with \code{scaling}. By scaling the
degree count by the total number of past events times two (i.e., use
\code{prop}), the statistic refers to the fraction of past events times two
that at least one actor in the pair was involved in. In this case, at the
first timepoint - when no events did previously occur - it is assumed that
every actor is equally likely to be involved in an event and the statistic
is set equal to 2/n, where n refers to the number of actors.
}
